# WMS_WebAPI/WMS_WebAPI/swagger/swaggerjson - Copy.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1"
  },
  "host": "localhost:50191",
  "schemes": [ "http", "https" ],
  "paths": {
...
    "/api/Warehouse/{id}": {
      "get": {
        "tags": [ "Warehouse" ],
        "operationId": "Warehouse_GetWarehouseByID",
        "consumes": [],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/Warehouse/userWarehouse": {
      "get": {
        "tags": [ "Warehouse" ],
        "operationId": "Warehouse_GetUserWarehouses",
        "consumes": [],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "userid",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "companyId",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
    "/api/Warehouse/Warehouse_Insert_Update": {
      "post": {
        "tags": [ "Warehouse" ],
        "operationId": "Warehouse_Warehouse_Insert_Update",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/cls_Warehouse" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "cls_Warehouse": {
      "type": "object",
      "properties": {
        "WarehouseID": {
          "format": "int32",
          "type": "integer"
        },
        "CompanyID": {
          "format": "int32",
          "type": "integer"
        },
        "WarehouseName": { "type": "string" },
        "WarehouseCode": { "type": "string" },
        "Address1": { "type": "string" },
        "Address2": { "type": "string" },
        "Address3": { "type": "string" },
        "TelNumber": { "type": "string" },
        "Fax": { "type": "string" },
        "EmailID": { "type": "string" },
        "CityId": {
          "format": "int32",
          "type": "integer"
        },
        "Logo": {
          "format": "byte",
          "type": "string"
        },
        "createdby": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  }
}